Reorganize packages
* direct_by_leaf/ is original StateDB package
* sql/ for SQL DB interfaces
