+++ direct_by_leaf/statedb.go
@@ -1,4 +1,4 @@
-package ipld_eth_statedb
+package state
 
 import (
 	"fmt"
