+++ sql/interfaces.go
@@ -1,14 +1,11 @@
-package ipld_eth_statedb
+package sql
 
 import (
 	"context"
 )
 
 // Database interfaces to support multiple Postgres drivers
-type Database interface {
-	Driver
-	Statements
-}
+type Database = Driver
 
 // Driver interface has all the methods required by a driver implementation to support the sql indexer
 type Driver interface {
@@ -24,11 +21,4 @@ type ScannableRow interface {
 // Result interface to accommodate different concrete result types
 type Result interface {
 	RowsAffected() (int64, error)
-}
-
-// Statements interface to accommodate different SQL query syntax
-type Statements interface {
-	GetContractCodeStmt() string
-	GetStateAccountStmt() string
-	GetStorageSlotStmt() string
 }
